* avoid a slice where stop < start (crashes after translation). * make sure we catch mallocs of negative size.

+++ pypy/translator/c/src/mem.h
@@ -9,7 +9,7 @@
 #define MAXIMUM_MALLOCABLE_SIZE   (LONG_MAX-4096)
 
 #define OP_MAX_VARSIZE(numitems, itemtype, err)  {			\
-    if ((numitems) > (MAXIMUM_MALLOCABLE_SIZE / sizeof(itemtype)))	\
+    if (((unsigned)(numitems)) > (MAXIMUM_MALLOCABLE_SIZE / sizeof(itemtype)))\
         FAIL_EXCEPTION(err, PyExc_MemoryError, "addr space overflow");	\
   }
